feat: event arrow
